minecraft_autonomous_builder/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minecraft-autonomous-builder"
version = "0.1.0"
description = "Dual-agent autonomous Minecraft building pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13",
  "fastapi>=0.111",
  "httpx>=0.27",
  "mcschematic>=11.4",
  "networkx>=3.3",
  "numpy>=1.26",
  "orjson>=3.10",
  "pydantic>=2.7",
  "PyYAML>=6.0.1",
  "tenacity>=8.3",
  "uvicorn>=0.30",
  "Pillow>=10.3",
  "requests>=2.32",
  "jsonschema>=4.22"
]

[project.optional-dependencies]
dev = [
  "black>=24.0",
  "ruff>=0.5",
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["bot", "data"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501", "E402", "W292", "I001"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q --cov=src --cov-report=term-missing"
